+++ CHANGELOG.md
@@ -463,3 +463,5 @@
 * Implemented backtick option for string sub literal strings
 * Initial resolution can now auto-set to native instead of fixed (behind CoreParams.SetNativeResolutionOnFirstRun flag)
 * Implemented PseudoExtensibleEnum groundwork, with tests and custom property drawer (this is one line, but it's huge!)
+* Added missing graphics to DefaultTheme
+* DefaultTheme is now applied by default